+++ apps/passport-client/components/modals/SettingsModal.tsx
@@ -41,16 +41,6 @@ export function SettingsModal() {
           {hasSetupPassword ? "Change" : "Add"} Password
         </LinkButton>
         <Spacer h={16} />
-        <LinkButton
-          $primary={true}
-          to="/subscriptions"
-          onClick={() =>
-            dispatch({ type: "set-modal", modal: { modalType: "none" } })
-          }
-        >
-          Manage Subscriptions
-        </LinkButton>
-        <Spacer h={16} />
         <Button onClick={clearZupass} style="danger">
           Log Out
         </Button>