+++ apps/passport-client/components/shared/AppHeader.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from "react";
 import styled from "styled-components";
-import { useDispatch } from "../../src/appHooks";
+import { useDispatch, useSubscriptions } from "../../src/appHooks";
 import { AppState } from "../../src/state";
 import { CircleButton } from "../core/Button";
 import { icons } from "../icons";
@@ -26,6 +26,8 @@ function AppHeaderImpl({ children }: { children?: React.ReactNode }) {
     () => setModal({ modalType: "settings" }),
     [setModal]
   );
+  const openSubscriptions = useCallback(() => window.location.hash = "subscriptions", []);
+  const subscriptions = useSubscriptions();
 
   return (
     <AppHeaderWrap>
@@ -33,6 +35,16 @@ function AppHeaderImpl({ children }: { children?: React.ReactNode }) {
         <img draggable="false" src={icons.infoAccent} width={34} height={34} />
       </CircleButton>
       {children}
+      <CircleButton diameter={34} padding={8} onClick={openSubscriptions}>
+        {subscriptions.value.getAllErrors().size > 0 &&
+          <ErrorDotContainer><ErrorDot /></ErrorDotContainer>}
+        <img
+          title="Subscriptions"
+          draggable="false"
+          src={icons.subscription}
+          width={34}
+          height={34} />
+      </CircleButton>
       <CircleButton diameter={34} padding={8} onClick={openSettings}>
         <img
           draggable="false"
@@ -52,4 +64,18 @@ const AppHeaderWrap = styled.div`
   justify-content: space-between;
   align-items: center;
   gap: 16px;
+`;
+
+const ErrorDot = styled.div`
+  position: absolute;
+  background: var(--danger);
+  right: 0px;
+  width: 8px;
+  height: 8px;
+  border-radius: 8px;
+`;
+
+const ErrorDotContainer = styled.div`
+  position: relative;
+  pointer-events: none;
 `;